refactor: do not use references to HTTP client
Finding issues when running it in the examples due to moving structs around while they hold references to an HTTP client which is also used elsewhere ("returns a value referencing data owned by the current function").

Currently I just `clone`, but we should probably have a blanket implementation for `Arc<C: HttpClient>` or something similar so we get to re-use the clients.