feat: use full filepath for key creation config
The `--name` passed as option to `create-system-identity` is the name used for the created private key in the current implementation. If no name is passed the key cannot be created, as the parent directory passed as `--output-local-path` is selected as the full filename. Besides, if the name contained spaces or other special characters it can cause problems when interpolating strings during operations like the Helm templating and other scripts.

Hence, we make `--output-local-path` specify the full filename to the private key and the name is not needed.